// src/services/youtube/@types/types.ts
export interface RetryOptions {
  maxRetries: number;
  backoff: {
    inc: number;
    max: number;
  };
}

export type PipelineEndpoint = (...args: any[]) => Promise<any>;

export interface RequestOptions {
  family?: number;
  localAddress?: string;
  headers: {
    Cookie?: string;
    cookie?: string;
    "x-youtube-client-name"?: string;
    "x-youtube-client-version"?: string;
    "x-youtube-identity-token"?: string | null;
  };
}

export interface InfoOptions {
  IPv6Block?: string;
  lang?: string;
  version?: string;
  requestCallback?: (req: any) => {};
  requestOptions?: {
    headers?: Record<string, string>;
  };
}

export type ExtendString<T extends string> = T | Omit<string, T>;
// ...
export type VideoFormatQuality =
  | "tiny"
  | "small"
  | "medium"
  | "large"
  | "hd720"
  | "hd1080"
  | "hd1440"
  | "hd2160"
  | "highres";

export interface VideoFormat {
  itag: number;
  url: string;
  mimeType?: string;
  bitrate?: number;
  audioBitrate?: number;
  width?: number;
  height?: number;
  initRange?: { start: string; end: string };
  indexRange?: { start: string; end: string };
  lastModified: string;
  contentLength: string;
  quality: ExtendString<VideoFormatQuality>;
  qualityLabel:
    | "144p"
    | "144p 15fps"
    | "144p60 HDR"
    | "240p"
    | "240p60 HDR"
    | "270p"
    | "360p"
    | "360p60 HDR"
    | "480p"
    | "480p60 HDR"
    | "720p"
    | "720p60"
    | "720p60 HDR"
    | "1080p"
    | "1080p60"
    | "1080p60 HDR"
    | "1440p"
    | "1440p60"
    | "1440p60 HDR"
    | "2160p"
    | "2160p60"
    | "2160p60 HDR"
    | "4320p"
    | "4320p60";
  projectionType?: "RECTANGULAR";
  fps?: number;
  averageBitrate?: number;
  audioQuality?:
    | "AUDIO_QUALITY_LOW"
    | "AUDIO_QUALITY_MEDIUM"
    | "AUDIO_QUALITY_HIGH";
  colorInfo?: {
    primaries: string;
    transferCharacteristics: string;
    matrixCoefficients: string;
  };
  highReplication?: boolean;
  approxDurationMs?: string;
  targetDurationSec?: number;
  maxDvrDurationSec?: number;
  audioSampleRate?: string;
  audioChannels?: number;

  container: string | null | "flv" | "3gp" | "mp4" | "webm" | "ts";
  hasVideo: boolean;
  hasAudio: boolean;
  codecs: string | null;
  videoCodec?: string | null;
  audioCodec?: string | null;

  isLive: boolean;
  isHLS: boolean;
  isDashMPD: boolean;
}

export interface thumbnail {
  url: string;
  width: number;
  height: number;
}
// ...
export interface captionTrack {
  baseUrl: string;
  name: {
    simpleText: ExtendString<CaptionTrackSimpleText>;
  };
  vssId: string;
  languageCode: ExtendString<CaptionTrackLanguageCode>;
  kind: string;
  rtl?: boolean;
  isTranslatable: boolean;
}

export interface audioTrack {
  captionTrackIndices: number[];
}

export interface translationLanguage {
  languageCode: captionTrack["languageCode"];
  languageName: captionTrack["name"];
}

export interface VideoDetails {
  videoId: string;
  title: string;
  shortDescription: string;
  lengthSeconds: string;
  keywords?: string[];
  channelId: string;
  isOwnerViewing: boolean;
  isCrawlable: boolean;
  thumbnails: thumbnail[];
  averageRating: number;
  allowRatings: boolean;
  viewCount: string;
  author: string;
  isPrivate: boolean;
  isUnpluggedCorpus: boolean;
  isLiveContent: boolean;
}

export interface Media {
  category: string;
  category_url: string;
  game?: string;
  game_url?: string;
  year?: number;
  song?: string;
  artist?: string;
  artist_url?: string;
  writers?: string;
  licensed_by?: string;
  thumbnails: thumbnail[];
}

export interface Author {
  id: string;
  name: string;
  avatar: string; // to remove later
  thumbnails?: thumbnail[];
  verified: boolean;
  user?: string;
  channel_url: string;
  external_channel_url?: string;
  user_url?: string;
  subscriber_count?: number;
}

export interface MicroformatRenderer {
  thumbnail: {
    thumbnails: thumbnail[];
  };
  embed: {
    iframeUrl: string;
    flashUrl: string;
    width: number;
    height: number;
    flashSecureUrl: string;
  };
  title: {
    simpleText: string;
  };
  description: {
    simpleText: string;
  };
  lengthSeconds: string;
  ownerProfileUrl: string;
  ownerGplusProfileUrl?: string;
  externalChannelId: string;
  isFamilySafe: boolean;
  availableCountries: string[];
  isUnlisted: boolean;
  hasYpcMetadata: boolean;
  viewCount: string;
  category: string;
  publishDate: string;
  ownerChannelName: string;
  liveBroadcastDetails?: {
    isLiveNow: boolean;
    startTimestamp: string;
    endTimestamp?: string;
  };
  uploadDate: string;
}

export interface storyboard {
  templateUrl: string;
  thumbnailWidth: number;
  thumbnailHeight: number;
  thumbnailCount: number;
  interval: number;
  columns: number;
  rows: number;
  storyboardCount: number;
}

export interface Chapter {
  title: string;
  start_time: number;
}

export interface MoreVideoDetails
  extends Omit<VideoDetails, "author" | "thumbnail" | "shortDescription">,
    Omit<MicroformatRenderer, "title" | "description"> {
  published: number;
  video_url: string;
  age_restricted: boolean;
  likes: number | null;
  dislikes: number | null;
  media: Media;
  author: Author;
  thumbnails: thumbnail[];
  storyboards: storyboard[];
  chapters: Chapter[];
  description: string | null;
}

export interface VideoInfo {
  iv_load_policy?: string;
  iv_allow_in_place_switch?: string;
  iv_endscreen_url?: string;
  iv_invideo_url?: string;
  iv3_module?: string;
  rmktEnabled?: string;
  uid?: string;
  vid?: string;
  focEnabled?: string;
  baseUrl?: string;
  storyboard_spec?: string;
  serialized_ad_ux_config?: string;
  player_error_log_fraction?: string;
  sffb?: string;
  ldpj?: string;
  videostats_playback_base_url?: string;
  innertube_context_client_version?: string;
  t?: string;
  fade_in_start_milliseconds: string;
  timestamp: string;
  ad3_module: string;
  relative_loudness: string;
  allow_below_the_player_companion: string;
  eventid: string;
  token: string;
  atc: string;
  cr: string;
  apply_fade_on_midrolls: string;
  cl: string;
  fexp: string[];
  apiary_host: string;
  fade_in_duration_milliseconds: string;
  fflags: string;
  ssl: string;
  pltype: string;
  enabled_engage_types: string;
  hl: string;
  is_listed: string;
  gut_tag: string;
  apiary_host_firstparty: string;
  enablecsi: string;
  csn: string;
  status: string;
  afv_ad_tag: string;
  idpj: string;
  sfw_player_response: string;
  account_playback_token: string;
  encoded_ad_safety_reason: string;
  tag_for_children_directed: string;
  no_get_video_log: string;
  ppv_remarketing_url: string;
  fmt_list: string[][];
  ad_slots: string;
  fade_out_duration_milliseconds: string;
  instream_long: string;
  allow_html5_ads: string;
  core_dbp: string;
  ad_device: string;
  itct: string;
  root_ve_type: string;
  excluded_ads: string;
  aftv: string;
  loeid: string;
  cver: string;
  shortform: string;
  dclk: string;
  csi_page_type: string;
  ismb: string;
  gpt_migration: string;
  loudness: string;
  ad_tag: string;
  of: string;
  probe_url: string;
  vm: string;
  afv_ad_tag_restricted_to_instream: string;
  gapi_hint_params: string;
  cid: string;
  c: string;
  oid: string;
  ptchn: string;
  as_launched_in_country: string;
  avg_rating: string;
  fade_out_start_milliseconds: string;
  midroll_prefetch_size: string;
  allow_ratings: string;
  thumbnail_url: string;
  iurlsd: string;
  iurlmq: string;
  iurlhq: string;
  iurlmaxres: string;
  ad_preroll: string;
  tmi: string;
  trueview: string;
  host_language: string;
  innertube_api_key: string;
  show_content_thumbnail: string;
  afv_instream_max: string;
  innertube_api_version: string;
  mpvid: string;
  allow_embed: string;
  ucid: string;
  plid: string;
  midroll_freqcap: string;
  ad_logging_flag: string;
  ptk: string;
  vmap: string;
  watermark: string[];
  dbp: string;
  ad_flags: string;
  html5player: string;
  formats: VideoFormat[];
  related_videos: relatedVideo[];
  no_embed_allowed?: boolean;
  player_response: PlayerResponse;
  videoDetails: MoreVideoDetails;
}

export interface PlayerResponse {
  playabilityStatus: {
    reason?: string;
    messages?: string[];
    status: string;
    errorScreen?: any;
    playableInEmbed: boolean;
    miniplayer: {
      miniplayerRenderer: {
        playbackMode: string;
      };
    };
    contextParams: string;
  };
  streamingData: {
    expiresInSeconds: string;
    formats: VideoFormat[];
    adaptiveFormats: VideoFormat[];
  };
  captions?: {
    playerCaptionsRenderer: {
      baseUrl: string;
      visibility: string;
    };
    playerCaptionsTracklistRenderer: {
      captionTracks: captionTrack[];
      audioTracks: audioTrack[];
      translationLanguages: translationLanguage[];
      defaultAudioTrackIndex: number;
    };
  };
  microformat: {
    playerMicroformatRenderer: MicroformatRenderer;
  };
  videoDetails: VideoDetails;
  playerConfig: {
    audioConfig: {
      loudnessDb: number;
      perceptualLoudnessDb: number;
      enablePerFormatLoudness: boolean;
    };
    streamSelectionConfig: { maxBitrate: string };
    mediaCommonConfig: { dynamicReadaheadConfig: {}[] };
    webPlayerConfig: { webPlayerActionsPorting: {}[] };
  };
}

export interface relatedVideo {
  id?: string;
  title?: string;
  published?: string;
  author: Author | "string"; // to remove the `string` part later
  ucid?: string; // to remove later
  author_thumbnail?: string; // to remove later
  short_view_count_text?: string;
  view_count?: string;
  length_seconds?: number;
  video_thumbnail?: string; // to remove later
  thumbnails: thumbnail[];
  richThumbnails: thumbnail[];
  isLive: boolean;
}